+++ esp-idf-svc/src/bt.rs
@@ -24,6 +24,7 @@ extern crate alloc;
 pub mod a2dp;
 #[cfg(all(esp32, esp_idf_bt_classic_enabled, esp_idf_bt_a2dp_enable))]
 pub mod avrc;
+#[cfg(esp_idf_bt_ble_enabled)]
 pub mod ble;
 #[cfg(all(esp32, esp_idf_bt_classic_enabled))]
 pub mod gap;